refactor: throw NoChanges at context
